requirements.txt
fastapi == 0.104.1
azure-cognitiveservices-speech
openai 
SpeechRecognition 
asyncio